+++ RobWorkSim/src/rwsim/util/SurfacePoseSampler.hpp
@@ -120,7 +120,7 @@ public:
 private:
     int binSearchRec(const double value, size_t start, size_t end){
         if(start==end)
-            return start;
+            return (int)start;
         // choose a int between start and end
         size_t split = (end-start)/2+start;
         if(value<_surfaceArea[split])
